compute cable resistance only when area or length are changed [...]
The `Cable` class used to calculateresistance (`r`) as a property, meaning
it recalculated the resistance every time the property was accessed.
It is more efficient to calculate it once when the `length` or `area` is
changed and store it as an attribute.
The `Cable` class now calculates and stores the resistance (`_r`) when the `length` or `area` is changed.