Revert "compute cable resistance only when area or length are changed [...]"
This reverts commit 9f6be7d1e65a5764bd0a5c7fa890407ef26806c9.

The reason is that cable area, length and resistivity should be already defined to compute
the resistance. The reverted commit introduced an error (thank you chatGPT...)
because it tried to calculate cable resistance before its area was defined
I'm lazyto fix it in a better way.

FYI the reason for the commit in the first place was to avoid to recompute r
at each call, but it's fast and not called that much, so should be fine...